pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Satinath Mondal"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.14"}
psycopg = {extras = ["binary"], version = "^3.2.9"}
fastmcp = "^2.10.1"
openai = "^1.93.0"
python-dotenv = "^1.1.1"
pydantic-settings = "^2.10.1"
httpx = "^0.28.1"
asyncio-mqtt = "^0.16.2"
gitpython = "^3.1.44"
pymongo = "^4.13.2"
motor = "^3.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
